Refactor and enable Cobalt browsertest execution on Android
Rename ContentShellBrowserTestActivity to CobaltBrowserTestActivity to accurately reflect Cobalt's standalone embedder architecture.
Implement native test command-line argument extraction in CobaltBrowserTestActivity using Chromium's existing NativeTestIntent constants.
Configure build targets and test launchers for Android browsertest execution.

BUG=b/338274026
TAG=agy
CONV=006ddb93-1731-4dba-b801-1f6fb67f1fd4
