+++ cobalt/testing/browser_tests/browsertests_apk/src/org/chromium/content_browsertests_apk/ContentBrowserTestsActivity.java
@@ -15,29 +15,25 @@
 package org.chromium.content_browsertests_apk;
 
 import android.os.Bundle;
-
 import java.io.File;
-
 import org.chromium.base.test.util.UrlUtils;
 
-/**
- * Android activity for running content_public.browser.tests
- */
-public class ContentBrowserTestsActivity extends ContentShellBrowserTestActivity {
-    private static final String TAG = "native_test";
+/** Android activity for running content_public.browser.tests */
+public class ContentBrowserTestsActivity extends CobaltBrowserTestActivity {
+  private static final String TAG = "native_test";
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        appendCommandLineFlags(
-                "--remote-debugging-socket-name content_browsertests_devtools_remote");
-    }
+  @Override
+  public void onCreate(Bundle savedInstanceState) {
+    appendCommandLineFlags("--remote-debugging-socket-name content_browsertests_devtools_remote");
+    super.onCreate(savedInstanceState);
+  }
 
-    @Override
-    protected File getPrivateDataDirectory() {
-        // TODO(agrieve): We should not be touching the side-loaded test data directory.
-        //     https://crbug.com/617734
-        return new File(UrlUtils.getIsolatedTestRoot(),
-                ContentBrowserTestsApplication.PRIVATE_DATA_DIRECTORY_SUFFIX);
-    }
+  @Override
+  protected File getPrivateDataDirectory() {
+    // TODO(agrieve): We should not be touching the side-loaded test data directory.
+    //     https://crbug.com/617734
+    return new File(
+        UrlUtils.getIsolatedTestRoot(),
+        ContentBrowserTestsApplication.PRIVATE_DATA_DIRECTORY_SUFFIX);
+  }
 }
